requirements.txt
openai
langchain
langchain_community
langchain_openai
langchainhub
bs4
pypdf
python-dotenv
faiss-cpu
fastapi
uvicorn
boto3
PyPDF2
langchain_chroma
FlagEmbedding